test 1

+++ src/main/java/com/tl_connect/dev/modules/academic_result/dto/SemesterSummaryDTO.java
@@ -1,5 +1,7 @@
 package com.tl_connect.dev.modules.academic_result.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -14,7 +16,7 @@ import lombok.Setter;
 public class SemesterSummaryDTO {
     private Integer creditsRegistered;
     private Integer creditsPassed;
-    private Double semesterGpa;
+    private BigDecimal semesterGpa;
     private Integer conductScore;
-    private Double cumulativeGpa;
+    private BigDecimal cumulativeGpa;
 }
