+++ src/main/java/com/tl_connect/dev/modules/academic_result/projection/SubjectResultRow.java
@@ -1,12 +1,14 @@
 package com.tl_connect.dev.modules.academic_result.projection;
 
+import java.math.BigDecimal;
+
 public interface SubjectResultRow {
     String getSemester();
     String getSubjectCode();
     String getSubjectName();
     Integer getCredits();
-    Double getScore10();
-    Double getScore4();
+    BigDecimal getScore10();
+    BigDecimal getScore4();
     String getLetterGrade();
     Boolean getIsPass();
 }
